Add Repository impl.
Remove ReadUpdateRepository
